Create read_pressure_psi.py
This example uses the Adafruit CircuitPython library for the ADS1x15 ADC and assumes you have it installed (pip install adafruit-circuitpython-ads1x15). Also, you need to connect the pressure sensor to the appropriate pins on your microcontroller.

Make sure to refer to the datasheets of your pressure sensor and ADC for accurate calibration information and connection details. The code above is just a starting point, and it needs to be adapted to your specific hardware setup.